chore: change rules of linter

+++ eslint.config.js
@@ -32,7 +32,7 @@ export default tseslint.config(
       ],
       'quotes': [
         'warn',
-        'single',
+        'double',
         {
           avoidEscape: true
         }
